refactor(runtime): runtime.go split — policy/runctx/toolpolicy/prompt files + provenance to journal (Phase 3.1 A-C)
Same-package splits along existing seams: policy.go (policyHook +
approval bundles), runctx.go (24 context keys + With*/fromCtx),
toolpolicy.go (tool-set shaping), prompt.go (8 pure injectors + persona
helpers + the RunWith injection block extracted as buildRunPrompt /
injectHostEnvironment — NOT merged across the buildLoopConfig boundary;
extracted bodies byte-identical to the originals). Why/Causes/ParentOf
move to kernel/journal/provenance.go as (j *Journal) methods (pure
Range scans; runtime keeps one-line delegators, controlplane untouched);
causation tests move with them. runtime.go 4,233 → 2,524.
