+++ kernel/runtime/runtime.go
@@ -25,7 +25,6 @@ import (
 	"maps"
 	"os"
 	"path/filepath"
-	stdruntime "runtime"
 	"slices"
 	"strings"
 	"sync"
@@ -41,7 +40,6 @@ import (
 	"github.com/agezt/agezt/kernel/cadence"
 	"github.com/agezt/agezt/kernel/catalog"
 	"github.com/agezt/agezt/kernel/configcenter"
-	"github.com/agezt/agezt/kernel/contextselect"
 	"github.com/agezt/agezt/kernel/datalake"
 	"github.com/agezt/agezt/kernel/edict"
 	"github.com/agezt/agezt/kernel/event"
@@ -1542,1169 +1540,6 @@ func (k *Kernel) RunPlan(ctx context.Context, plan scheduler.Plan, planID string
 	return k.scheduler.Run(runCtx, plan, planID)
 }
 
-// policyHook adapts the kernel's Edict engine to the agent.Policy
-// signature the tool-loop expects. It is called once per ToolCall,
-// before invocation.
-//
-// Three paths:
-//
-//  1. Hard-deny / unknown-cap / L0 / AskDeny → Allow=false, run skipped.
-//  2. L4 Allow or AskAllow folded Ask → Allow=true.
-//  3. AskPrompt landed on Ask-class → submit to approval.Registry and
-//     block on the operator's decision. Grant flips Allow=true; deny /
-//     timeout / cancel keep Allow=false with the verdict reason.
-//
-// The ctx passed in is the per-run context; cancellation (Halt) flows
-// through to Submit and surfaces as DecisionCancel.
-// validatedToolCaps keeps only declarations naming a capability Edict knows
-// (M900) — a plugin may join an existing policy axis, never invent one.
-func validatedToolCaps(declared map[string]string) map[string]edict.Capability {
-	if len(declared) == 0 {
-		return nil
-	}
-	out := make(map[string]edict.Capability, len(declared))
-	for tool, cap := range declared {
-		if edict.KnownCapability(cap) {
-			out[tool] = edict.Capability(cap)
-		}
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-func (k *Kernel) policyHook(ctx context.Context, tc agent.ToolCall) agent.PolicyVerdict {
-	// Declared-capability overlay (M900): a plugin tool whose manifest joined
-	// a known policy axis is classified there; everything else falls through
-	// to the built-in name/input classification.
-	cap, declared := k.toolCaps[tc.Name]
-	if !declared {
-		cap = edict.CapabilityForToolCall(tc.Name, tc.Input)
-	}
-	var out edict.Outcome
-	if ceiling, ok := trustCeilingFromCtx(ctx); ok {
-		out = k.edict.DecideWithCeiling(cap, string(tc.Input), ceiling) // SPEC-16 §4 initiative ceiling
-	} else {
-		out = k.edict.Decide(cap, string(tc.Input))
-	}
-
-	verdict := agent.PolicyVerdict{
-		Allow:      out.Decision == edict.DecisionAllow,
-		Capability: string(out.Capability),
-		Reason:     out.Reason,
-		WouldAsk:   out.WouldAsk,
-		HardDenied: out.HardDenied,
-	}
-	def, _ := agent.PolicyToolDefFromContext(ctx)
-	bundle := k.approvalDecisionBundle(tc.Name, out.Capability, tc.Input, def)
-	verdict.EffectClass = bundle.EffectClass
-	verdict.AffectedResources = append([]string(nil), bundle.AffectedResources...)
-	ep := k.epistemicGate(tc.Name, out.Capability, tc.Input, def, bundle)
-	verdict.EpistemicAction = ep.Action
-	verdict.EpistemicReason = ep.Reason
-	verdict.EpistemicSignals = append([]string(nil), ep.Signals...)
-	verdict.EpistemicConfidence = ep.Confidence
-	verdict.FailureMatches = ep.FailureMatches
-	verdict.WeightedFailures = ep.WeightedFailures
-	verdict.SchemaHash = ep.SchemaHash
-	verdict.InputShape = ep.InputShape
-	verdict.TemporalSensitive = ep.Temporal
-	verdict.NovelTool = ep.NovelTool
-	if reason, denied := agentToolPolicyDenial(agentToolPolicyFromCtx(ctx), tc.Name); denied {
-		verdict.Allow = false
-		verdict.Reason = reason
-		verdict.WouldAsk = false
-		verdict.HardDenied = true
-		return verdict
-	}
-	if reason, denied := k.agentNoisePolicyDenial(ctx, tc); denied {
-		verdict.Allow = false
-		verdict.Reason = reason
-		verdict.WouldAsk = false
-		verdict.HardDenied = true
-		return verdict
-	}
-	taint, hasTaint := agent.UntrustedObservationTaintFromContext(ctx)
-	if hasTaint {
-		verdict.UntrustedObservation = true
-		verdict.ObservationSources = append([]string(nil), taint.Sources...)
-		verdict.ObservationDirectiveLike = taint.DirectiveLike
-		verdict.ObservationDirectiveMatches = append([]string(nil), taint.Matches...)
-	}
-	intentFrame, hasIntentFrame := intentmodel.FrameFromContext(ctx)
-	intentAction := intentmodel.Action{
-		ToolName:          tc.Name,
-		Capability:        string(out.Capability),
-		EffectClass:       bundle.EffectClass,
-		Input:             string(tc.Input),
-		AffectedResources: append([]string(nil), bundle.AffectedResources...),
-	}
-	regretAxes := intentmodel.RegretForAction(intentAction)
-	confirmationPrompt := ""
-	if hasIntentFrame {
-		confirmationPrompt = intentmodel.ConfirmationPrompt(intentFrame, intentAction, regretAxes)
-	}
-
-	requiresApproval := out.RequiresApproval
-	approvalReason := out.Reason
-	if k.cfg.EpistemicEscalation && verdict.Allow && ep.escalates() {
-		requiresApproval = true
-		approvalReason = ep.Reason
-		verdict.Allow = false
-		verdict.WouldAsk = true
-	}
-	if k.cfg.IntentRegretGating && verdict.Allow && hasIntentFrame && intentmodel.RequiresConfirmation(intentFrame, regretAxes) {
-		requiresApproval = true
-		approvalReason = confirmationPrompt
-		verdict.Allow = false
-		verdict.WouldAsk = true
-		k.publishIntentConfirmationRequired(correlationFromCtx(ctx), actorFromCtx(ctx), intentFrame, regretAxes, confirmationPrompt)
-	}
-	// Prompt-injection guard: an effectful action within the causal window of a
-	// directive-like untrusted observation. The agent loop already scoped
-	// taint.DirectiveLike to that window, so this no longer fires for the whole
-	// run after one suspicious observation.
-	if k.cfg.PromptInjectionGuard != PromptInjectionOff && verdict.Allow && hasTaint && taint.DirectiveLike && bundle.EffectClass != string(agent.EffectReadOnly) {
-		// Block only in On mode and only when the operator hasn't trusted this
-		// run; warn mode and a trusted run audit without interrupting.
-		if k.cfg.PromptInjectionGuard == PromptInjectionOn && !trustedObservations(ctx) {
-			requiresApproval = true
-			approvalReason = "prompt-injection guard: effectful action is downstream of directive-like untrusted observation from " + strings.Join(taint.Sources, ", ")
-			verdict.Allow = false
-			verdict.WouldAsk = true
-		} else {
-			k.publishPromptInjectionWarned(correlationFromCtx(ctx), actorFromCtx(ctx), tc.Name, string(out.Capability), taint.Sources, trustedObservations(ctx))
-		}
-	}
-
-	// Session-scoped operator grant (chat "auto-approve Tool Forge this session"):
-	// if the run carries an auto-approve set covering this capability, satisfy the
-	// approval without prompting and journal it as an auto-grant (WouldAsk stays
-	// true so `agt why` shows it would have asked). Hard-denies never reach here
-	// (they resolve to deny, not approval), so this can't override the F4 floor.
-	if requiresApproval && autoApproveCap(ctx, string(out.Capability)) {
-		verdict.Allow = true
-		verdict.WouldAsk = true
-		k.publishAutoApprove(correlationFromCtx(ctx), actorFromCtx(ctx), string(out.Capability), tc.Name)
-		return verdict
-	}
-
-	if !requiresApproval {
-		return verdict
-	}
-
-	// Live HITL: pause the tool-loop, route the request through the
-	// approval queue, block until decided.
-	actor := actorFromCtx(ctx)
-	corr := correlationFromCtx(ctx)
-	res := k.approvals.Submit(ctx, approval.SubmitSpec{
-		Capability:            string(out.Capability),
-		ToolName:              tc.Name,
-		Input:                 string(tc.Input),
-		Reason:                approvalReason,
-		Actor:                 actor,
-		CorrelationID:         corr,
-		EffectClass:           bundle.EffectClass,
-		PredictedEffects:      bundle.PredictedEffects,
-		AffectedResources:     bundle.AffectedResources,
-		RollbackNotes:         bundle.RollbackNotes,
-		Confidence:            bundle.Confidence,
-		CanonicalIntent:       intentFrame.CanonicalIntent,
-		HarmfulInterpretation: intentFrame.HarmfulReading,
-		AmbiguityScore:        intentFrame.AmbiguityScore,
-		RegretAxes:            regretAxesPayload(regretAxes),
-		ConfirmationPrompt:    confirmationPrompt,
-	})
-	switch res.Decision {
-	case approval.DecisionGrant:
-		verdict.Allow = true
-		verdict.Reason = "approval granted by " + res.ResolvedBy
-	default:
-		verdict.Allow = false
-		verdict.Reason = fmt.Sprintf("approval %s: %s", res.Decision, res.Reason)
-	}
-	return verdict
-}
-
-func (k *Kernel) agentNoisePolicyDenial(ctx context.Context, tc agent.ToolCall) (string, bool) {
-	policy, ok := agentNoisePolicyFromCtx(ctx)
-	if !ok {
-		return "", false
-	}
-	if tc.Name == "memory" && policy.disableMemoryWrites && memoryToolActionWrites(tc.Input) {
-		return "agent noise policy: memory writes are disabled", true
-	}
-	if tc.Name != "notify" {
-		return "", false
-	}
-	if min := notifySeverityRank(policy.minNotifySeverity); min > notifySeverityRank(notifySeverityFromInput(tc.Input)) {
-		return fmt.Sprintf("agent noise policy: notify severity must be at least %s", policy.minNotifySeverity), true
-	}
-	if policy.minNotifyIntervalSec <= 0 {
-		return "", false
-	}
-	slug, _ := agentIdentFromCtx(ctx)
-	if strings.TrimSpace(slug) == "" {
-		return "", false
-	}
-	nowMS := time.Now().UnixMilli()
-	var st agentNoiseState
-	if raw, ok, err := k.state.Get(agentNoiseStateNS, slug); err != nil {
-		return "agent noise policy: notify cooldown state unavailable: " + err.Error(), true
-	} else if ok {
-		_ = json.Unmarshal(raw, &st)
-	}
-	if st.PendingNotifyMS > 0 && nowMS-st.PendingNotifyMS < int64(agentNoisePendingNotifyTTL/time.Millisecond) {
-		return "agent noise policy: notify send already in progress", true
-	}
-	if st.LastNotifyMS > 0 {
-		elapsed := nowMS - st.LastNotifyMS
-		minMS := int64(policy.minNotifyIntervalSec) * 1000
-		if elapsed < minMS {
-			remaining := time.Duration(minMS-elapsed) * time.Millisecond
-			return "agent noise policy: notify cooldown active for " + remaining.Round(time.Second).String(), true
-		}
-	}
-	st.PendingNotifyMS = nowMS
-	if err := k.state.Set(agentNoiseStateNS, slug, st); err != nil {
-		return "agent noise policy: notify cooldown state unavailable: " + err.Error(), true
-	}
-	return "", false
-}
-
-func (k *Kernel) completeAgentNoiseNotify(ctx context.Context, tc agent.ToolCall, res agent.Result) {
-	policy, ok := agentNoisePolicyFromCtx(ctx)
-	if !ok || policy.minNotifyIntervalSec <= 0 || tc.Name != "notify" {
-		return
-	}
-	slug, _ := agentIdentFromCtx(ctx)
-	if strings.TrimSpace(slug) == "" {
-		return
-	}
-	var st agentNoiseState
-	if raw, ok, err := k.state.Get(agentNoiseStateNS, slug); err != nil {
-		return
-	} else if ok {
-		_ = json.Unmarshal(raw, &st)
-	}
-	st.PendingNotifyMS = 0
-	if !res.IsError {
-		st.LastNotifyMS = time.Now().UnixMilli()
-	}
-	_ = k.state.Set(agentNoiseStateNS, slug, st)
-}
-
-func memoryToolActionWrites(raw json.RawMessage) bool {
-	var in struct {
-		Action string `json:"action"`
-	}
-	if json.Unmarshal(raw, &in) != nil {
-		return false
-	}
-	switch strings.ToLower(strings.TrimSpace(in.Action)) {
-	case "remember", "forget", "bulk_forget":
-		return true
-	default:
-		return false
-	}
-}
-
-func notifySeverityFromInput(raw json.RawMessage) string {
-	var in struct {
-		Severity string `json:"severity"`
-	}
-	if json.Unmarshal(raw, &in) != nil {
-		return "info"
-	}
-	severity := strings.ToLower(strings.TrimSpace(in.Severity))
-	if severity == "" {
-		return "info"
-	}
-	return severity
-}
-
-func notifySeverityRank(severity string) int {
-	switch strings.ToLower(strings.TrimSpace(severity)) {
-	case "critical":
-		return 3
-	case "warning", "warn":
-		return 2
-	case "info", "":
-		return 1
-	default:
-		return 0
-	}
-}
-
-type approvalBundle struct {
-	EffectClass       string
-	PredictedEffects  []string
-	AffectedResources []string
-	RollbackNotes     string
-	Confidence        float64
-}
-
-func (k *Kernel) approvalDecisionBundle(toolName string, cap edict.Capability, input json.RawMessage, def agent.ToolDef) approvalBundle {
-	effect := def.Effect
-	if effect.Class == "" && len(effect.PredictedEffects) == 0 && len(effect.AffectedResources) == 0 {
-		if tool, ok := k.tools[toolName]; ok {
-			effect = tool.Definition().Effect
-		}
-	}
-
-	class := normalizeEffectClass(effect.Class)
-	if class == "" {
-		class = defaultEffectClass(cap)
-	}
-	resources := append([]string(nil), effect.AffectedResources...)
-	if len(resources) == 0 {
-		resources = affectedResourcesFromInput(toolName, cap, input)
-	}
-	predicted := append([]string(nil), effect.PredictedEffects...)
-	if len(predicted) == 0 {
-		predicted = []string{fmt.Sprintf("invoke %s under %s", toolName, cap)}
-	}
-	rollback := strings.TrimSpace(effect.RollbackNotes)
-	if rollback == "" {
-		rollback = defaultRollbackNotes(class)
-	}
-	confidence := effect.Confidence
-	if confidence <= 0 || confidence > 1 {
-		confidence = defaultEffectConfidence(class)
-	}
-	return approvalBundle{
-		EffectClass:       class,
-		PredictedEffects:  predicted,
-		AffectedResources: resources,
-		RollbackNotes:     rollback,
-		Confidence:        confidence,
-	}
-}
-
-func normalizeEffectClass(class agent.EffectClass) string {
-	switch class {
-	case agent.EffectReadOnly, agent.EffectReversible, agent.EffectCompensable, agent.EffectIrreversible:
-		return string(class)
-	default:
-		return ""
-	}
-}
-
-func defaultEffectClass(cap edict.Capability) string {
-	switch cap {
-	case edict.CapFileRead, edict.CapFileList, edict.CapHTTPGet, edict.CapBrowserRead,
-		edict.CapHomeAssistantRead, edict.CapWebSearch, edict.CapRunsRead,
-		edict.CapIntrospect, edict.CapConfigRead, edict.CapProviderCall:
-		return string(agent.EffectReadOnly)
-	case edict.CapFileWrite, edict.CapMemory, edict.CapWorld, edict.CapSchedule,
-		edict.CapStanding, edict.CapBoard, edict.CapSkill, edict.CapOversee,
-		edict.CapToolForge, edict.CapConfigWrite, edict.CapWorkflow:
-		return string(agent.EffectReversible)
-	case edict.CapNotify, edict.CapHTTPPost, edict.CapRemoteRun:
-		return string(agent.EffectCompensable)
-	case edict.CapShell, edict.CapFileDelete, edict.CapCoding, edict.CapACPAgent,
-		edict.CapHomeAssistantCall, edict.CapCodeExec, edict.CapMCPInstall, edict.CapMCP:
-		return string(agent.EffectIrreversible)
-	default:
-		return string(agent.EffectIrreversible)
-	}
-}
-
-func affectedResourcesFromInput(toolName string, cap edict.Capability, input json.RawMessage) []string {
-	out := []string{"tool:" + toolName, "capability:" + string(cap)}
-	var obj map[string]any
-	if err := json.Unmarshal(input, &obj); err != nil {
-		return out
-	}
-	for _, key := range []string{"path", "url", "endpoint", "entity_id", "service", "command", "op", "name", "id"} {
-		if v, ok := obj[key]; ok {
-			if s, ok := v.(string); ok && strings.TrimSpace(s) != "" {
-				out = append(out, key+":"+s)
-			}
-		}
-	}
-	return out
-}
-
-func defaultRollbackNotes(class string) string {
-	switch class {
-	case string(agent.EffectReadOnly):
-		return "No rollback required for read-only action."
-	case string(agent.EffectReversible):
-		return "Use the corresponding revert/delete/restore operation or journaled state to undo if needed."
-	case string(agent.EffectCompensable):
-		return "No guaranteed rollback; compensate with a follow-up action if the outcome is wrong."
-	case string(agent.EffectIrreversible):
-		return "No reliable rollback path declared; approve only if the effect is acceptable."
-	default:
-		return "No rollback information declared."
-	}
-}
-
-func defaultEffectConfidence(class string) float64 {
-	switch class {
-	case string(agent.EffectReadOnly):
-		return 0.95
-	case string(agent.EffectReversible):
-		return 0.75
-	case string(agent.EffectCompensable):
-		return 0.6
-	case string(agent.EffectIrreversible):
-		return 0.5
-	default:
-		return 0.4
-	}
-}
-
-// per-run context keys used by RunWith → policyHook to carry the
-// actor/correlation IDs into approval.Submit so audit events stay
-// linked to the originating task.
-type ctxKey int
-
-const (
-	ctxKeyActor ctxKey = iota
-	ctxKeyCorrelation
-	ctxKeyModel
-	ctxKeyImages
-	ctxKeySystem
-	ctxKeyRunTimeout
-	ctxKeyTools
-	ctxKeyMaxCost
-	ctxKeyJSONMode
-	ctxKeyTrustCeiling
-	ctxKeyRoot
-	ctxKeyModelChain
-	ctxKeyAgentIdent
-	ctxKeySystemAgent
-	ctxKeyAgentLifecycle
-	ctxKeyAgentRetryPolicy
-	ctxKeyAgentToolPolicy
-	ctxKeyAgentConfigOverrides
-	ctxKeyAgentNoisePolicy
-	ctxKeyWakeContext
-	ctxKeyAutoApproveCaps
-	ctxKeyTrustedObservations
-	ctxKeyResumeOwned // M1002: a resume ticket for this corr is already owned by an outer frame
-	ctxKeyResumeSeed  // M1002: prior conversation + iter to seed a resumed run
-)
-
-// agentIdent carries a named agent's identity + daily ceiling for the
-// Governor's per-agent ledger (M793).
-type agentIdent struct {
-	slug    string
-	dailyMc int64
-}
-
-type agentToolPolicy struct {
-	allow []string
-	deny  []string
-}
-
-const agentNoiseStateNS = "agent_noise"
-
-type agentNoisePolicy struct {
-	silentOnSuccess      bool
-	disableMemoryWrites  bool
-	minNotifySeverity    string
-	minNotifyIntervalSec int
-}
-
-type agentNoiseState struct {
-	LastNotifyMS    int64 `json:"last_notify_ms"`
-	PendingNotifyMS int64 `json:"pending_notify_ms,omitempty"`
-}
-
-const agentNoisePendingNotifyTTL = 5 * time.Minute
-
-// WakeContext is durable provenance for why a run exists. It is stamped on
-// task.received by agent.Run and intentionally kept separate from the prompt.
-type WakeContext struct {
-	Source            string
-	Reason            string
-	ScheduleID        string
-	StandingID        string
-	StandingName      string
-	TriggerSubject    string
-	ParentCorrelation string
-}
-
-func cloneStringMap(in map[string]string) map[string]string {
-	if len(in) == 0 {
-		return nil
-	}
-	out := make(map[string]string, len(in))
-	for key, value := range in {
-		out[key] = value
-	}
-	return out
-}
-
-// rootFromCtx returns the correlation of the ROOT run of a delegation tree (the
-// top-level lead), propagated to every descendant so a tree-wide cap can be
-// attributed to the whole tree rather than a single spawner. Empty when not in
-// a delegated context.
-func rootFromCtx(ctx context.Context) string {
-	if v, ok := ctx.Value(ctxKeyRoot).(string); ok {
-		return v
-	}
-	return ""
-}
-
-// WithTrustCeiling returns a context capping autonomous tool-use at `ceiling` for
-// the run started with it (SPEC-16 §4 initiative.max_trust). The policy hook
-// consults it so a normally auto-allowed capability is downgraded to Ask (or
-// Deny) within this run. ceiling >= LevelAllow is a no-op (no clamp). Used by the
-// standing-order runner to bound an order's autonomy.
-//
-// The ceiling is monotonically TIGHTENING down a delegation tree: if the context
-// already carries a tighter (lower) ceiling, that one is kept. A child run (e.g. a
-// delegated sub-agent whose profile declares a looser TrustCeiling) can therefore
-// never loosen the bound its parent was started with — only narrow it. Without
-// this, WithAgentProfile re-applying a target profile's higher ceiling would
-// overwrite a standing-order initiative cap and let delegation escape it
-// (CWE-269, security finding VULN-001).
-func WithTrustCeiling(ctx context.Context, ceiling edict.TrustLevel) context.Context {
-	if ceiling >= edict.LevelAllow {
-		// "No clamp" must not erase an existing tighter ceiling: leave ctx as-is so
-		// any inherited cap survives.
-		return ctx
-	}
-	if existing, ok := trustCeilingFromCtx(ctx); ok && existing < ceiling {
-		ceiling = existing
-	}
-	return context.WithValue(ctx, ctxKeyTrustCeiling, ceiling)
-}
-
-func trustCeilingFromCtx(ctx context.Context) (edict.TrustLevel, bool) {
-	v, ok := ctx.Value(ctxKeyTrustCeiling).(edict.TrustLevel)
-	return v, ok
-}
-
-// WithImages returns a context carrying image-attachment references for the run
-// started with it (M93). They flow into the agent loop's initial user message.
-// Empty is a no-op. The caller (control plane) only sets this after the M91
-// vision gate confirms the active model is vision-capable.
-func WithImages(ctx context.Context, images []string) context.Context {
-	if len(images) == 0 {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyImages, images)
-}
-
-func imagesFromCtx(ctx context.Context) []string {
-	if v, ok := ctx.Value(ctxKeyImages).([]string); ok {
-		return v
-	}
-	return nil
-}
-
-// WithJSONMode returns a context requesting structured (JSON) output for the run
-// started with it (M314). It flows into the agent loop's CompletionRequest.JSONMode,
-// so a provider with a native JSON mode constrains its output. false is a no-op.
-// Used by the OpenAI-compatible API to honour a client's response_format.
-func WithJSONMode(ctx context.Context, jsonMode bool) context.Context {
-	if !jsonMode {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyJSONMode, true)
-}
-
-func jsonModeFromCtx(ctx context.Context) bool {
-	v, _ := ctx.Value(ctxKeyJSONMode).(bool)
-	return v
-}
-
-// WithModel returns a context that overrides the model for the run started with
-// it. Empty model is a no-op (the kernel's configured Model is used). The
-// override flows into the agent loop's CompletionRequest.Model, so the selected
-// provider serves exactly the requested model — the basis for per-request model
-// selection from the OpenAI-compatible API.
-func WithModel(ctx context.Context, model string) context.Context {
-	if model == "" {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyModel, model)
-}
-
-func modelFromCtx(ctx context.Context) string {
-	if v, ok := ctx.Value(ctxKeyModel).(string); ok {
-		return v
-	}
-	return ""
-}
-
-// WithSystem returns a context that overrides the base system prompt for the run
-// started with it (M148-sibling). Empty is a no-op (the kernel's configured System
-// is used). The override REPLACES the configured System; memory/world/skill
-// injection still layer on top, so a one-off identity/instruction override can
-// be set per run without losing what Agezt knows.
-func WithSystem(ctx context.Context, system string) context.Context {
-	if system == "" {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeySystem, system)
-}
-
-func systemFromCtx(ctx context.Context) string {
-	if v, ok := ctx.Value(ctxKeySystem).(string); ok {
-		return v
-	}
-	return ""
-}
-
-// WithRunTimeout returns a context that overrides the per-run wall-clock budget
-// for the run started with it (a per-run counterpart to Config.MaxDuration / M31).
-// d <= 0 is a no-op (the configured MaxDuration, if any, applies). Lets a single
-// run be bounded without a daemon-wide cap (`agt run --timeout`).
-func WithRunTimeout(ctx context.Context, d time.Duration) context.Context {
-	if d <= 0 {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyRunTimeout, d)
-}
-
-func runTimeoutFromCtx(ctx context.Context) time.Duration {
-	if v, ok := ctx.Value(ctxKeyRunTimeout).(time.Duration); ok {
-		return v
-	}
-	return 0
-}
-
-// WithMaxCost returns a context that caps the cumulative provider spend (in
-// USD-microcents) for the run started with it (M166) — the per-run cost analogue
-// of WithRunTimeout. mc <= 0 is a no-op (uncapped). Lets a single run be bounded
-// by money (`agt run --max-cost`) without a daemon-wide ceiling; the Governor's
-// daily ceiling still applies on top.
-func WithMaxCost(ctx context.Context, mc int64) context.Context {
-	if mc <= 0 {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyMaxCost, mc)
-}
-
-func maxCostFromCtx(ctx context.Context) int64 {
-	if v, ok := ctx.Value(ctxKeyMaxCost).(int64); ok {
-		return v
-	}
-	return 0
-}
-
-// WithAutoApproveCapabilities marks a set of capabilities to auto-grant when
-// the policy would otherwise prompt for HITL approval, for THIS run and every
-// sub-agent it spawns (the context value rides the delegation tree). caps is a
-// set of edict capability strings (e.g. {"tool.forge","code.exec"}). This is a
-// session-scoped operator grant — e.g. the chat "auto-approve Tool Forge for
-// this session" toggle when standing up an agent army — NOT a daemon-wide policy
-// change. It never overrides a hard-deny (those resolve to deny, not approval).
-// Empty leaves the context unchanged.
-func WithAutoApproveCapabilities(ctx context.Context, caps map[string]bool) context.Context {
-	if len(caps) == 0 {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyAutoApproveCaps, caps)
-}
-
-func mergeAutoApproveCapabilities(ctx context.Context, caps map[string]bool) map[string]bool {
-	if len(caps) == 0 {
-		return nil
-	}
-	out := make(map[string]bool, len(caps))
-	if existing, ok := ctx.Value(ctxKeyAutoApproveCaps).(map[string]bool); ok {
-		for c, on := range existing {
-			if on {
-				out[c] = true
-			}
-		}
-	}
-	for c, on := range caps {
-		if on {
-			out[c] = true
-		}
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-// autoApproveCap reports whether capability c is in this run's auto-approve set.
-func autoApproveCap(ctx context.Context, c string) bool {
-	if v, ok := ctx.Value(ctxKeyAutoApproveCaps).(map[string]bool); ok {
-		return v[c]
-	}
-	return false
-}
-
-// PromptInjectionMode selects how the prompt-injection guard handles an
-// effectful action downstream of directive-like untrusted content.
-type PromptInjectionMode int
-
-const (
-	// PromptInjectionOn routes the action to HITL approval.
-	PromptInjectionOn PromptInjectionMode = iota
-	// PromptInjectionWarn allows the action but journals a warning (default).
-	PromptInjectionWarn
-	// PromptInjectionOff disables the active intervention entirely.
-	PromptInjectionOff
-)
-
-// ParsePromptInjectionMode maps an operator string (AGEZT_PROMPT_INJECTION_GUARD)
-// to a mode. "off"/"0"/"false" → Off, "warn"/"warning"/"audit" → Warn, anything
-// else (including empty) → Warn. Operators who want a live approval stop on
-// directive-like untrusted content set "on"/"block"/"prompt".
-func ParsePromptInjectionMode(s string) PromptInjectionMode {
-	switch strings.ToLower(strings.TrimSpace(s)) {
-	case "off", "0", "false", "no":
-		return PromptInjectionOff
-	case "on", "1", "true", "yes", "block", "prompt":
-		return PromptInjectionOn
-	case "", "warn", "warning", "audit":
-		return PromptInjectionWarn
-	default:
-		return PromptInjectionWarn
-	}
-}
-
-// WithTrustedObservations marks a run as one whose untrusted-observation content
-// the operator has chosen to trust (e.g. the chat "trust this run's web content"
-// toggle). It downgrades the prompt-injection guard from blocking to warn FOR
-// THIS RUN and its sub-agents, so a deliberately operator-driven agentic task
-// isn't interrupted for every action — without changing the daemon-wide posture
-// or touching any hard-deny. Never affects the F4 floor or SSRF/budget guards.
-func WithTrustedObservations(ctx context.Context) context.Context {
-	return context.WithValue(ctx, ctxKeyTrustedObservations, true)
-}
-
-// trustedObservations reports whether this run carries the operator's
-// trust-this-run grant.
-func trustedObservations(ctx context.Context) bool {
-	v, _ := ctx.Value(ctxKeyTrustedObservations).(bool)
-	return v
-}
-
-// WithAgentProfile applies a roster profile to a run's context (M790): the
-// soul becomes the system override, the model + ordered fallbacks become the
-// run's model chain, and the memory scope follows the identity (M786). The
-// per-run cost ceiling is NOT applied here — callers layer it so their own
-// explicit budget wins (mirrors handleRun's precedence). Used by the standing
-// runner so an order can fire AS a named agent; handleRun keeps its inline
-// application (its model resolves before the vision gate).
-func WithAgentProfile(ctx context.Context, p roster.Profile) context.Context {
-	noise := effectiveAgentNoisePolicy(p)
-	if p.System {
-		ctx = context.WithValue(ctx, ctxKeySystemAgent, true)
-	}
-	if sys := agentProfileSystem(p); sys != "" {
-		ctx = WithSystem(ctx, sys)
-	}
-	if p.Lifecycle.Mode != "" || p.Lifecycle.RetireOnComplete {
-		ctx = context.WithValue(ctx, ctxKeyAgentLifecycle, p.Lifecycle)
-	}
-	if p.RetryPolicy != nil {
-		cp := *p.RetryPolicy
-		cp.RetryOn = append([]string(nil), p.RetryPolicy.RetryOn...)
-		ctx = context.WithValue(ctx, ctxKeyAgentRetryPolicy, cp)
-	}
-	if len(p.ToolAllow) > 0 || len(p.ToolDeny) > 0 {
-		ctx = context.WithValue(ctx, ctxKeyAgentToolPolicy, agentToolPolicy{
-			allow: append([]string(nil), p.ToolAllow...),
-			deny:  append([]string(nil), p.ToolDeny...),
-		})
-	}
-	if noise != (agentNoisePolicy{}) {
-		ctx = context.WithValue(ctx, ctxKeyAgentNoisePolicy, noise)
-	}
-	if len(p.ConfigOverrides) > 0 {
-		ctx = context.WithValue(ctx, ctxKeyAgentConfigOverrides, cloneStringMap(p.ConfigOverrides))
-	}
-	if ceiling := strings.TrimSpace(p.TrustCeiling); ceiling != "" {
-		if lvl, err := edict.ParseTrustLevel(ceiling); err == nil {
-			ctx = WithTrustCeiling(ctx, lvl)
-		}
-	}
-	primary := strings.TrimSpace(p.Model)
-	if primary != "" {
-		ctx = WithModel(ctx, primary)
-	}
-	if len(p.Fallbacks) > 0 {
-		chain := []string{primary}
-		if primary == "" {
-			chain = nil
-		}
-		for _, m := range p.Fallbacks {
-			if m = strings.TrimSpace(m); m != "" && m != primary {
-				chain = append(chain, m)
-			}
-		}
-		ctx = WithModelChain(ctx, chain)
-	}
-	scope := strings.TrimSpace(p.MemoryScope)
-	if scope == "" {
-		scope = p.Slug
-	}
-	ctx = memory.WithScope(ctx, scope)
-	// The agent's working directory (M792): file/shell tools operate inside
-	// this workspace subdirectory. Escape-proofed by the setter.
-	ctx = agent.WithWorkdir(ctx, p.Workdir)
-	// And its identity + daily ceiling for the Governor's ledger (M793).
-	return WithAgentIdent(ctx, p.Slug, p.MaxDailyMc)
-}
-
-func effectiveAgentNoisePolicy(p roster.Profile) agentNoisePolicy {
-	var out agentNoisePolicy
-	if p.NoisePolicy != nil {
-		out.silentOnSuccess = p.NoisePolicy.SilentOnSuccess
-		out.disableMemoryWrites = p.NoisePolicy.DisableMemoryWrites
-		out.minNotifySeverity = strings.ToLower(strings.TrimSpace(p.NoisePolicy.MinNotifySeverity))
-		out.minNotifyIntervalSec = p.NoisePolicy.MinNotifyIntervalSec
-	}
-	if out.silentOnSuccess && notifySeverityRank(out.minNotifySeverity) < notifySeverityRank("warning") {
-		out.minNotifySeverity = "warning"
-	}
-	if p.System {
-		out.silentOnSuccess = true
-		out.disableMemoryWrites = true
-		if notifySeverityRank(out.minNotifySeverity) < notifySeverityRank("warning") {
-			out.minNotifySeverity = "warning"
-		}
-		if out.minNotifyIntervalSec < 8*3600 {
-			out.minNotifyIntervalSec = 8 * 3600
-		}
-	}
-	return out
-}
-
-func agentNoisePolicyFromCtx(ctx context.Context) (agentNoisePolicy, bool) {
-	v, ok := ctx.Value(ctxKeyAgentNoisePolicy).(agentNoisePolicy)
-	return v, ok
-}
-
-func applyAgentNoisePolicyToPromptTools(tools map[string]agent.Tool, ctx context.Context) map[string]agent.Tool {
-	policy, ok := agentNoisePolicyFromCtx(ctx)
-	if !ok || !policy.disableMemoryWrites {
-		return tools
-	}
-	if _, ok := tools["memory"]; !ok {
-		return tools
-	}
-	out := make(map[string]agent.Tool, len(tools)-1)
-	for name, tool := range tools {
-		if name != "memory" {
-			out[name] = tool
-		}
-	}
-	return out
-}
-
-func appendUniqueString(in []string, value string) []string {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return in
-	}
-	for _, x := range in {
-		if strings.EqualFold(strings.TrimSpace(x), value) {
-			return in
-		}
-	}
-	return append(in, value)
-}
-
-// WithAgentIdent stamps the run with a named agent's identity and per-day
-// spend ceiling (M793): every completion of the run is metered against the
-// Governor's per-agent daily ledger and refused past the ceiling.
-func WithAgentIdent(ctx context.Context, slug string, dailyMc int64) context.Context {
-	if strings.TrimSpace(slug) == "" {
-		return ctx
-	}
-	// Also stamp the agent slug under the kernel/agent key so provenance-aware
-	// tools (memory, M851) can read who is acting via agent.AgentFromContext —
-	// the runtime key here is private and additionally carries the daily ceiling.
-	ctx = agent.WithAgent(ctx, slug)
-	return context.WithValue(ctx, ctxKeyAgentIdent, agentIdent{slug: slug, dailyMc: dailyMc})
-}
-
-func agentIdentFromCtx(ctx context.Context) (string, int64) {
-	if v, ok := ctx.Value(ctxKeyAgentIdent).(agentIdent); ok {
-		return v.slug, v.dailyMc
-	}
-	return "", 0
-}
-
-func agentSlugFromCtx(ctx context.Context) string { s, _ := agentIdentFromCtx(ctx); return s }
-
-func agentDailyMcFromCtx(ctx context.Context) int64 { _, d := agentIdentFromCtx(ctx); return d }
-
-// WithWakeContext attaches run provenance to the next agent loop. Empty fields
-// are omitted from the journal; callers can layer it with WithAgentProfile.
-func WithWakeContext(ctx context.Context, w WakeContext) context.Context {
-	w.Source = strings.TrimSpace(w.Source)
-	w.Reason = strings.TrimSpace(w.Reason)
-	w.ScheduleID = strings.TrimSpace(w.ScheduleID)
-	w.StandingID = strings.TrimSpace(w.StandingID)
-	w.StandingName = strings.TrimSpace(w.StandingName)
-	w.TriggerSubject = strings.TrimSpace(w.TriggerSubject)
-	w.ParentCorrelation = strings.TrimSpace(w.ParentCorrelation)
-	if w == (WakeContext{}) {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyWakeContext, w)
-}
-
-func wakeContextFromCtx(ctx context.Context) WakeContext {
-	v, _ := ctx.Value(ctxKeyWakeContext).(WakeContext)
-	return v
-}
-
-func systemAgentFromCtx(ctx context.Context) bool {
-	v, _ := ctx.Value(ctxKeySystemAgent).(bool)
-	return v
-}
-
-func agentToolPolicyFromCtx(ctx context.Context) agentToolPolicy {
-	v, _ := ctx.Value(ctxKeyAgentToolPolicy).(agentToolPolicy)
-	return v
-}
-
-func agentRetryPolicyFromCtx(ctx context.Context) (roster.RetryPolicy, bool) {
-	v, ok := ctx.Value(ctxKeyAgentRetryPolicy).(roster.RetryPolicy)
-	if !ok {
-		return roster.RetryPolicy{}, false
-	}
-	v.RetryOn = append([]string(nil), v.RetryOn...)
-	return v, true
-}
-
-// AgentConfigOverrides returns the named agent's config-override map attached to
-// a run context by WithAgentProfile. The returned map is a copy and safe for the
-// caller to mutate. Nil means the run carries no agent-specific overrides.
-func AgentConfigOverrides(ctx context.Context) map[string]string {
-	v, _ := ctx.Value(ctxKeyAgentConfigOverrides).(map[string]string)
-	return cloneStringMap(v)
-}
-
-func agentConfigStringOverride(ctx context.Context, key string) (string, bool) {
-	v, ok := agentConfigOverrideRaw(AgentConfigOverrides(ctx), key)
-	if !ok {
-		return "", false
-	}
-	return agentConfigStringValue(v)
-}
-
-func agentConfigBoolOverride(ctx context.Context, key string) (bool, bool) {
-	raw, ok := agentConfigStringOverride(ctx, key)
-	if !ok {
-		return false, false
-	}
-	return agentConfigBoolValue(raw)
-}
-
-func agentConfigIntOverride(ctx context.Context, key string) (int, bool) {
-	raw, ok := agentConfigStringOverride(ctx, key)
-	if !ok {
-		return 0, false
-	}
-	return agentConfigIntValue(raw)
-}
-
-func agentConfigDurationOverride(ctx context.Context, key string) (time.Duration, bool) {
-	raw, ok := agentConfigStringOverride(ctx, key)
-	if !ok {
-		return 0, false
-	}
-	return agentConfigDurationValue(raw)
-}
-
-func agentProfileSystem(p roster.Profile) string {
-	var b strings.Builder
-	if soul := strings.TrimSpace(p.Soul); soul != "" {
-		b.WriteString(soul)
-	}
-	if len(p.Instructions) > 0 {
-		if b.Len() > 0 {
-			b.WriteString("\n\n")
-		}
-		b.WriteString("Standing instructions:\n")
-		for _, ins := range p.Instructions {
-			if ins = strings.TrimSpace(ins); ins != "" {
-				b.WriteString("- ")
-				b.WriteString(ins)
-				b.WriteString("\n")
-			}
-		}
-	}
-	if len(p.TaskList) > 0 {
-		cycle, total := profileTasksByScope(p.TaskList)
-		if len(cycle) > 0 || len(total) > 0 {
-			if b.Len() > 0 {
-				b.WriteString("\n")
-			}
-			if len(cycle) > 0 {
-				b.WriteString("\nCycle tasks:\n")
-				writeProfileTasks(&b, cycle)
-			}
-			if len(total) > 0 {
-				b.WriteString("\nTotal tasks:\n")
-				writeProfileTasks(&b, total)
-			}
-		}
-	}
-	return strings.TrimSpace(b.String())
-}
-
-func profileTasksByScope(tasks []roster.AgentTask) (cycle, total []roster.AgentTask) {
-	for _, t := range tasks {
-		status := strings.TrimSpace(t.Status)
-		if status == "done" || status == "retired" {
-			continue
-		}
-		if strings.TrimSpace(t.Scope) == "cycle" {
-			cycle = append(cycle, t)
-		} else {
-			total = append(total, t)
-		}
-	}
-	return cycle, total
-}
-
-func writeProfileTasks(b *strings.Builder, tasks []roster.AgentTask) {
-	for i, t := range tasks {
-		if i >= 20 {
-			b.WriteString("- ...\n")
-			return
-		}
-		title := strings.TrimSpace(t.Title)
-		if title == "" {
-			continue
-		}
-		status := strings.TrimSpace(t.Status)
-		if status == "" {
-			status = "todo"
-		}
-		b.WriteString("- [")
-		b.WriteString(status)
-		b.WriteString("] ")
-		b.WriteString(title)
-		if desc := strings.TrimSpace(t.Description); desc != "" {
-			b.WriteString(" - ")
-			b.WriteString(desc)
-		}
-		b.WriteString("\n")
-	}
-}
-
-// WithModelChain sets the run's per-agent ordered model fallback chain (M787):
-// the Governor tries these models in order, overriding the task type's
-// configured chain. Carries a named agent's own fallbacks (roster M783).
-func WithModelChain(ctx context.Context, chain []string) context.Context {
-	if len(chain) == 0 {
-		return ctx
-	}
-	return context.WithValue(ctx, ctxKeyModelChain, chain)
-}
-
-func modelChainFromCtx(ctx context.Context) []string {
-	if v, ok := ctx.Value(ctxKeyModelChain).([]string); ok {
-		return v
-	}
-	return nil
-}
-
-// WithTools restricts the run started with this context to the named tools only
-// (a per-run allowlist). A non-nil slice — including an EMPTY one (no tools at all,
-// for a pure-reasoning / safe one-off run) — activates the restriction; passing it
-// is the only way to override, so an unrestricted run is simply one where this is
-// never called. Names not registered are ignored.
-func WithTools(ctx context.Context, allow []string) context.Context {
-	return context.WithValue(ctx, ctxKeyTools, allow)
-}
-
-// toolsFromCtx returns the per-run tool allowlist and whether one was set. ok=false
-// means "no restriction" (use all tools); ok=true with an empty/nil slice means
-// "no tools".
-func toolsFromCtx(ctx context.Context) ([]string, bool) {
-	v, ok := ctx.Value(ctxKeyTools).([]string)
-	return v, ok
-}
-
-// filterTools returns the subset of tools whose names are in allow (a registered
-// name not present is dropped; an allow name with no matching tool is ignored).
-// An empty/nil allow yields an empty map — no tools.
-func filterTools(tools map[string]agent.Tool, allow []string) map[string]agent.Tool {
-	keep := make(map[string]struct{}, len(allow))
-	for _, n := range allow {
-		if n = strings.ToLower(strings.TrimSpace(n)); n != "" {
-			keep[n] = struct{}{}
-		}
-	}
-	out := make(map[string]agent.Tool, len(keep))
-	for name, tool := range tools {
-		if _, ok := keep[strings.ToLower(name)]; ok {
-			out[name] = tool
-		}
-	}
-	return out
-}
-
-func applyAgentToolPolicy(tools map[string]agent.Tool, pol agentToolPolicy) map[string]agent.Tool {
-	out := tools
-	if len(pol.allow) > 0 {
-		out = filterTools(out, pol.allow)
-	}
-	if len(pol.deny) == 0 {
-		return out
-	}
-	deny := make(map[string]struct{}, len(pol.deny))
-	for _, name := range pol.deny {
-		if name = strings.ToLower(strings.TrimSpace(name)); name != "" {
-			deny[name] = struct{}{}
-		}
-	}
-	next := make(map[string]agent.Tool, len(out))
-	for name, tool := range out {
-		if _, blocked := deny[strings.ToLower(name)]; blocked {
-			continue
-		}
-		next[name] = tool
-	}
-	return next
-}
-
-func agentToolPolicyDenial(pol agentToolPolicy, toolName string) (string, bool) {
-	name := strings.ToLower(strings.TrimSpace(toolName))
-	if name == "" {
-		return "", false
-	}
-	for _, denied := range pol.deny {
-		if strings.ToLower(strings.TrimSpace(denied)) == name {
-			return "agent tool denylist", true
-		}
-	}
-	if len(pol.allow) == 0 {
-		return "", false
-	}
-	for _, allowed := range pol.allow {
-		if strings.ToLower(strings.TrimSpace(allowed)) == name {
-			return "", false
-		}
-	}
-	return "not in agent tool allowlist", true
-}
-
-func actorFromCtx(ctx context.Context) string {
-	if v, ok := ctx.Value(ctxKeyActor).(string); ok {
-		return v
-	}
-	return ""
-}
-
-func correlationFromCtx(ctx context.Context) string {
-	if v, ok := ctx.Value(ctxKeyCorrelation).(string); ok {
-		return v
-	}
-	return ""
-}
-
 // IsHalted reports whether Run will refuse to start.
 func (k *Kernel) IsHalted() bool {
 	k.runsMu.Lock()
@@ -3238,176 +2073,9 @@ func (k *Kernel) RunWith(ctx context.Context, corr, intent string) (string, erro
 		}
 	}
 
-	// Memory injection: recall relevant records and prepend them to the
-	// system prompt so the model starts the task already knowing what
-	// Agezt remembers. The recall is journaled (memory.retrieved) under
-	// corr, so `agt why` shows exactly what knowledge was surfaced.
-	// Per-run system-prompt override (WithSystem): a one-off identity/instruction
-	// set for this run only; falls back to the kernel's configured System. Memory /
-	// world / skill injection below still layer on top.
-	system := k.System() // live daemon default identity (M710), editable at runtime
-	if s := systemFromCtx(runCtx); s != "" {
-		system = s
-	}
-	// Operator profile (M1000): prepend what AGEZT has learned about the operator
-	// so every (non-system) run knows WHO it works for — distinct from the per-agent
-	// persona above (identity) and applied before the ephemeral task injections
-	// below, so it sits adjacent to the persona. Always-on (not intent-driven);
-	// a no-op until DistillProfile has synthesized a profile.
-	if k.cfg.ProfileInject && !systemAgent {
-		if p := k.memory.ProfileText(); p != "" {
-			system = injectUserProfile(system, p)
-		}
-	}
-	// Taste overlay: prepend curated "what good looks like" exemplars scoped to
-	// this run so output quality is anchored to concrete examples. Sits adjacent
-	// to the operator profile (both shape HOW the agent works) and before the
-	// factual memory recall below. Journaled as taste.injected under corr.
-	if k.cfg.TasteInject && !systemAgent && k.taste != nil {
-		topK := k.cfg.TasteTopK
-		if topK <= 0 {
-			topK = 3
-		}
-		if ex := k.taste.ForScope(agentSlugFromCtx(runCtx), topK); len(ex) > 0 {
-			system = injectTaste(system, ex)
-			ids := make([]string, 0, len(ex))
-			for _, e := range ex {
-				ids = append(ids, e.ID)
-			}
-			_, _ = k.bus.Publish(event.Spec{
-				Subject:       "agent.agent-" + corr + ".taste",
-				Kind:          event.KindTasteInjected,
-				Actor:         "taste",
-				CorrelationID: corr,
-				Payload:       map[string]any{"count": len(ex), "ids": ids, "scope": agentSlugFromCtx(runCtx)},
-			})
-		}
-	}
-	if k.cfg.MemoryInject && !systemAgent {
-		topK := k.cfg.MemoryTopK
-		if topK <= 0 {
-			topK = 5
-		}
-		var candidates []contextselect.Candidate
-		if scored, err := k.memory.SearchScoped(intent, contextselect.CandidateLimit, memory.ScopeFrom(runCtx)); err == nil {
-			candidates = memoryContextCandidates(scored, time.Now().UnixMilli())
-		}
-		// Scoped to the run's agent identity (M786): a named agent's private
-		// notes surface in its injected context; an unscoped run sees shared
-		// memory only (RecallScoped with "" ≡ the previous Recall behaviour).
-		if hits, err := k.memory.RecallScoped(corr, intent, topK, memory.ScopeFrom(runCtx)); err == nil && len(hits) > 0 {
-			system = injectMemory(system, hits)
-			ids := make([]string, 0, len(hits))
-			for _, h := range hits {
-				ids = append(ids, h.Record.ID)
-			}
-			chosen, rejected := contextselect.SplitCandidates(candidates, contextselect.ChosenIDSet(ids), "memory_recall")
-			k.publishContextSelection(corr, actor, contextselect.Manifest{
-				Phase:    "memory",
-				Query:    intent,
-				Chosen:   chosen,
-				Rejected: rejected,
-				Summary:  contextselect.Summary(chosen, rejected),
-			})
-		}
-	}
-
-	// World-model injection: resolve the entities the intent refers to and
-	// prepend them, so the model starts knowing what "the portfolio" means
-	// (SPEC-05 §7 step 1). Resolve journals worldmodel.retrieved under corr,
-	// so `agt why` shows what references were grounded.
-	if k.cfg.WorldInject && !systemAgent {
-		topK := k.cfg.WorldTopK
-		if topK <= 0 {
-			topK = 5
-		}
-		var candidates []contextselect.Candidate
-		if scored, err := k.world.ResolveQuiet(intent, contextselect.CandidateLimit); err == nil {
-			candidates = worldContextCandidates(scored, time.Now().UnixMilli())
-		}
-		if hits, err := k.world.Resolve(corr, intent, topK); err == nil && len(hits) > 0 {
-			system = injectWorld(system, hits)
-			ids := make([]string, 0, len(hits))
-			for _, h := range hits {
-				ids = append(ids, h.Entity.ID)
-			}
-			chosen, rejected := contextselect.SplitCandidates(candidates, contextselect.ChosenIDSet(ids), "world_resolve")
-			k.publishContextSelection(corr, actor, contextselect.Manifest{
-				Phase:    "world",
-				Query:    intent,
-				Chosen:   chosen,
-				Rejected: rejected,
-				Summary:  contextselect.Summary(chosen, rejected),
-			})
-		}
-	}
-
-	// Skill activation: retrieve matching ACTIVE skills and prepend their
-	// bodies so the model plans with learned procedures (SPEC-05 §4.2, §7
-	// step 4). The pool is scoped to the acting agent (M932): shared skills
-	// plus its own private ones. Activate journals skill.activated under corr
-	// for `agt why`.
-	var activatedSkillIDs []string
-	if k.cfg.SkillInject && !systemAgent {
-		topK := k.cfg.SkillTopK
-		if topK <= 0 {
-			topK = 3
-		}
-		agentSlug := agentSlugFromCtx(runCtx)
-		if skillDirective.Explicit {
-			hits, missing, err := k.forge.ActivateExplicitFor(corr, agentSlug, intent, skillDirective.Refs, topK)
-			if err == nil {
-				if len(hits) > 0 {
-					system = injectSkills(system, hits)
-					for _, h := range hits {
-						activatedSkillIDs = append(activatedSkillIDs, h.Skill.ID)
-					}
-				}
-				chosen := skillContextCandidates(hits, time.Now().UnixMilli())
-				for i := range chosen {
-					chosen[i].Chosen = true
-					chosen[i].Reason = "selected:skill_explicit_activation"
-				}
-				summary := contextselect.Summary(chosen, nil)
-				summary["activation"] = "explicit"
-				summary["refs"] = skillDirective.Refs
-				if len(missing) > 0 {
-					summary["missing"] = missing
-				}
-				k.publishContextSelection(corr, actor, contextselect.Manifest{
-					Phase:   "skill",
-					Query:   intent,
-					Chosen:  chosen,
-					Summary: summary,
-				})
-			}
-		} else {
-			var candidates []contextselect.Candidate
-			if all, err := k.forge.List(); err == nil {
-				pool := all[:0:0]
-				for _, sk := range all {
-					if sk.Agent == "" || sk.Agent == agentSlug {
-						pool = append(pool, sk)
-					}
-				}
-				candidates = skillContextCandidates(skill.Retrieve(pool, intent, contextselect.CandidateLimit, time.Now().UnixMilli()), time.Now().UnixMilli())
-			}
-			if hits, err := k.forge.ActivateFor(corr, agentSlug, intent, topK); err == nil && len(hits) > 0 {
-				system = injectSkills(system, hits)
-				for _, h := range hits {
-					activatedSkillIDs = append(activatedSkillIDs, h.Skill.ID)
-				}
-				chosen, rejected := contextselect.SplitCandidates(candidates, contextselect.ChosenIDSet(activatedSkillIDs), "skill_activation")
-				k.publishContextSelection(corr, actor, contextselect.Manifest{
-					Phase:    "skill",
-					Query:    intent,
-					Chosen:   chosen,
-					Rejected: rejected,
-					Summary:  contextselect.Summary(chosen, rejected),
-				})
-			}
-		}
-	}
+	// System-prompt assembly: per-run override or live default, then
+	// profile/taste/memory/world/skill injection (buildRunPrompt).
+	system, activatedSkillIDs := k.buildRunPrompt(runCtx, corr, actor, intent, systemAgent, skillDirective)
 
 	// Per-run model override (WithModel) — used by the OpenAI-compatible API and
 	// the Chat picker to honour the request's `model`. Falls back to the live
@@ -3438,14 +2106,8 @@ func (k *Kernel) RunWith(ctx context.Context, corr, intent string) (string, erro
 	// accounting, compaction, or tool policy (LD-1).
 	lc := k.buildLoopConfig(runCtx, corr, model)
 
-	// Host-environment preamble (M609): prepend OS/arch, the shell the shell tool
-	// uses, the shared workspace dir, the date, and THIS run's tools — so the
-	// model acts correctly on this host instead of guessing. Injected last (after
-	// memory/world/skills and after the run's tool set is resolved) so it sits at
-	// the top of the system prompt and reflects any per-run tool restriction.
-	if k.cfg.EnvironmentInject {
-		system = injectEnvironment(system, k.cfg.WorkspaceRoot, lc.Tools, time.Now())
-	}
+	// Host-environment preamble (M609) — see injectHostEnvironment.
+	system = k.injectHostEnvironment(system, lc.Tools)
 
 	wake := wakeContextFromCtx(runCtx)
 
@@ -3753,274 +2415,6 @@ func makeElidedSummarizer(provider agent.Provider, model, corr string, maxTokens
 	}
 }
 
-// injectMemory prepends a compact "Relevant memory" block to the system
-// prompt. Records are rendered one per line as "- [TYPE] subject: content".
-func injectMemory(system string, hits []memory.Scored) string {
-	var b strings.Builder
-	b.WriteString("Relevant memory (recalled from prior tasks; use if helpful):\n")
-	for _, h := range hits {
-		fmt.Fprintf(&b, "- [%s] %s: %s\n", h.Record.Type, h.Record.Subject, h.Record.Content)
-	}
-	if system != "" {
-		b.WriteString("\n")
-		b.WriteString(system)
-	}
-	return b.String()
-}
-
-// injectUserProfile prepends the learned operator profile (M1000) to the system
-// prompt so the agent always knows who it works for. profileText is the
-// pre-formatted facet block from memory.ProfileText (never empty here).
-func injectUserProfile(system, profileText string) string {
-	var b strings.Builder
-	b.WriteString("What you know about the operator you work for (apply naturally; don't recite):\n")
-	b.WriteString(profileText)
-	b.WriteString("\n")
-	if system != "" {
-		b.WriteString("\n")
-		b.WriteString(system)
-	}
-	return b.String()
-}
-
-// injectTaste prepends a "what good looks like" block of curated exemplars to
-// the system prompt so the model anchors its output to concrete examples of good
-// work. Each exemplar renders as a titled block; scoped ones are already ordered
-// first by the store.
-func injectTaste(system string, exemplars []taste.Exemplar) string {
-	var b strings.Builder
-	b.WriteString("What good looks like (curated exemplars — match this quality and style, don't copy verbatim):\n")
-	for _, e := range exemplars {
-		b.WriteString("\n### ")
-		b.WriteString(e.Title)
-		b.WriteString("\n")
-		b.WriteString(e.Body)
-		b.WriteString("\n")
-	}
-	if system != "" {
-		b.WriteString("\n")
-		b.WriteString(system)
-	}
-	return b.String()
-}
-
-// injectWorld prepends a compact "Known entities" block to the system prompt.
-// Entities are rendered one per line as "- [kind] name (aliases: ...)" so the
-// model can ground references like "the portfolio" to concrete things.
-func injectWorld(system string, hits []worldmodel.ScoredEntity) string {
-	var b strings.Builder
-	b.WriteString("Known entities (from the world model; use to ground references):\n")
-	for _, h := range hits {
-		e := h.Entity
-		if len(e.Aliases) > 0 {
-			fmt.Fprintf(&b, "- [%s] %s (aka %s)\n", e.Kind, e.Name, strings.Join(e.Aliases, ", "))
-		} else {
-			fmt.Fprintf(&b, "- [%s] %s\n", e.Kind, e.Name)
-		}
-	}
-	if system != "" {
-		b.WriteString("\n")
-		b.WriteString(system)
-	}
-	return b.String()
-}
-
-// injectSkills prepends matching active skills' bodies to the system prompt so
-// the model plans with learned procedures. Each is rendered as a titled block.
-func injectSkills(system string, hits []skill.Scored) string {
-	var b strings.Builder
-	b.WriteString("Applicable skills (learned procedures; follow if relevant):\n")
-	for _, h := range hits {
-		s := h.Skill
-		fmt.Fprintf(&b, "## %s — %s\n%s\n", s.Name, s.Description, s.Body)
-		// A bundled skill (agentskills.io shape, M847) ships reference files and
-		// scripts. List them and tell the agent how to reach them: read a reference
-		// with `skill op=read`, run a script with shell/code_exec from the dir that
-		// `skill op=files` reports. This is what lets a skill say "run scripts/setup.sh
-		// to install the CLI" and have the agent actually do it.
-		if len(s.Resources) > 0 {
-			fmt.Fprintf(&b, "Bundled resources (use the `skill` tool: op=files for the directory, op=read \"<path>\" to read one; run scripts with shell/code_exec):\n")
-			for _, r := range s.Resources {
-				fmt.Fprintf(&b, "  - %s\n", r)
-			}
-		}
-	}
-	if system != "" {
-		b.WriteString("\n")
-		b.WriteString(system)
-	}
-	return b.String()
-}
-
-// shellHinter is the optional interface a shell-like tool implements to tell the
-// environment preamble the EXACT interpreter it runs (binary + flag), so the
-// guidance reflects an operator's shell override rather than a GOOS guess.
-type shellHinter interface{ ShellHint() (string, string) }
-
-// injectEnvironment prepends a concise host-environment preamble to the system
-// prompt (M609): OS/arch, the shell the shell tool uses (with command-style
-// guidance so the model doesn't try `ls` on Windows), the shared workspace dir,
-// the date, and the run's available tools. This is the single highest-leverage
-// fix for blind trial-and-error tool use on non-Unix hosts. `now` is passed in
-// for deterministic tests.
-func injectEnvironment(system, workspaceRoot string, tools map[string]agent.Tool, now time.Time) string {
-	var b strings.Builder
-	b.WriteString("## Runtime environment\n")
-	b.WriteString("You run on a real host — act for THIS environment, do not assume Unix.\n")
-	fmt.Fprintf(&b, "- OS / arch: %s / %s\n", stdruntime.GOOS, stdruntime.GOARCH)
-
-	// Shell line: prefer the shell tool's own hint (honours overrides); fall back
-	// to the GOOS default the shell tool would pick.
-	shellBin, shellArg := defaultShellHint()
-	if t, ok := tools["shell"]; ok {
-		if h, ok := t.(shellHinter); ok {
-			shellBin, shellArg = h.ShellHint()
-		}
-	}
-	fmt.Fprintf(&b, "- Shell tool runs commands via `%s %s`. %s\n", shellBin, shellArg, shellGuidance(shellBin))
-
-	if workspaceRoot != "" {
-		fmt.Fprintf(&b, "- Working directory (shell + file tools both operate here): %s\n", workspaceRoot)
-	}
-	fmt.Fprintf(&b, "- Today: %s\n", now.Format("2006-01-02"))
-
-	if len(tools) > 0 {
-		names := make([]string, 0, len(tools))
-		for name := range tools {
-			names = append(names, name)
-		}
-		slices.Sort(names)
-		b.WriteString("- Tools available this run:\n")
-		for _, name := range names {
-			fmt.Fprintf(&b, "  - %s — %s\n", name, firstSentence(tools[name].Definition().Description))
-		}
-	}
-	if brief := capabilityBriefing(tools); brief != "" {
-		b.WriteString(brief)
-	}
-	if bias := forgeBias(tools); bias != "" {
-		b.WriteString(bias)
-	}
-	b.WriteString("Some capabilities require operator approval and may be denied; if a call is denied, adapt your approach rather than repeating it.\n")
-
-	if system != "" {
-		b.WriteString("\n")
-		b.WriteString(system)
-	}
-	return b.String()
-}
-
-// capabilityBriefing tells the agent, plainly, how much it is allowed to do —
-// so it knows its own reach instead of timidly assuming limits (M848). AGEZT is
-// default-allow (owner law): every capability is on unless explicitly opted out,
-// so an agent should marshal whatever it takes to finish the task — write and run
-// code, install and run CLIs and packages, build whole apps and services. The
-// briefing is tuned to the tools ACTUALLY present this run (no point promising
-// code execution if code_exec was opted out), and stays honest about the few real
-// rails: explicit operator denials, budgets, and the SSRF/secret guards. Returns
-// "" when none of the relevant tools are available.
-func capabilityBriefing(tools map[string]agent.Tool) string {
-	_, hasShell := tools["shell"]
-	_, hasCode := tools["code_exec"]
-	_, hasFile := tools["file"]
-	_, hasForge := tools["tool_forge"]
-	_, hasSkill := tools["skill"]
-	if !hasShell && !hasCode && !hasFile {
-		return "" // nothing to brief — this run can't build or run anything
-	}
-	var b strings.Builder
-	b.WriteString("## What you can do — act without artificial limits\n")
-	b.WriteString("You are a full agent on a real machine, not a sandboxed chat. To accomplish the task you may do whatever it takes — there are no artificial capability limits. Be bold: build it, install it, run it.\n")
-	if hasCode {
-		b.WriteString("- Write and run real code — Python, Node/JavaScript, Deno — via code_exec. Build scripts, prototypes, data pipelines, whole programs.\n")
-	}
-	if hasShell {
-		b.WriteString("- Install and run anything the host supports via the shell: CLI tools, npm / pip / cargo / go packages, build systems, even long-running background services. If a command is missing, install it, then use it.\n")
-	}
-	if hasFile {
-		b.WriteString("- Create and edit as many files, projects, and applications as the task needs in your working directory.\n")
-	}
-	if hasForge {
-		b.WriteString("- When a one-off script isn't enough, forge your own durable tool (tool_forge) so the capability persists.\n")
-	}
-	if hasSkill {
-		b.WriteString("- Capture what works as a reusable skill — including bundled reference files and scripts — so future runs reuse it (skill op=learn / op=files / op=read).\n")
-	}
-	b.WriteString("Default to action: prefer doing the work over asking whether you're allowed. The only real limits are explicit — a denied approval, a spend budget, and the network/secret guards (no SSRF, secrets stay redacted). Everything else is yours to use.\n")
-	return b.String()
-}
-
-// forgeBias nudges the agent toward DETERMINISM and SELF-IMPROVEMENT (M902): for
-// work that must be exact, is repeatable, or you'll likely do again, prefer a
-// tool over ad-hoc reasoning — write a script so the result is deterministic and
-// re-runnable, forge a recurring script into a durable tool, and capture what
-// works as a skill so it compounds across runs. Tuned to the tools present;
-// returns "" when none of code_exec / tool_forge / skill is available (nothing
-// to bias toward). Complements capabilityBriefing (which says what you CAN do)
-// with how to work well.
-func forgeBias(tools map[string]agent.Tool) string {
-	_, hasCode := tools["code_exec"]
-	_, hasForge := tools["tool_forge"]
-	_, hasSkill := tools["skill"]
-	if !hasCode && !hasForge && !hasSkill {
-		return ""
-	}
-	var b strings.Builder
-	b.WriteString("## Prefer deterministic tools — and improve your own\n")
-	b.WriteString("For work that must be exact, is repeatable, or you'll likely do again, reach for a tool instead of reasoning it out by hand each time:\n")
-	if hasCode {
-		b.WriteString("- Write a script (code_exec) so the result is deterministic, checkable, and re-runnable — not re-derived (and error-prone) each turn. Computation, parsing, transforms, and anything with exact rules belong in code.\n")
-	}
-	if hasForge {
-		b.WriteString("- When a one-off script recurs, forge it into a durable tool (tool_forge) so the capability persists and the next run just calls it.\n")
-	}
-	if hasSkill {
-		b.WriteString("- Check existing skills/tools before re-deriving one, and capture a working approach as a reusable skill (skill op=learn) so it's there next time.\n")
-	}
-	b.WriteString("Treat each run as self-improvement: when you hit a capability gap, build the tool that closes it — you finish faster and more reliably every time after.\n")
-	return b.String()
-}
-
-// defaultShellHint mirrors the shell tool's platform default for callers that
-// can't reach the live tool (e.g. tools map without a shell). cmd on Windows,
-// sh elsewhere — kept in sync with plugins/tools/shell.resolveShell.
-func defaultShellHint() (string, string) {
-	if stdruntime.GOOS == "windows" {
-		return "cmd", "/C"
-	}
-	return "sh", "-c"
-}
-
-// shellGuidance returns one line of command-style advice keyed off the shell
-// binary, so the model uses native commands (the #1 source of wasted iterations
-// on Windows was the model reflexively trying `ls`/`cat`/`rm`).
-func shellGuidance(shellBin string) string {
-	// Normalize Windows separators first: on Linux filepath.Base treats
-	// `C:\Win\cmd.exe` as one element and the interpreter would misroute
-	// to the POSIX advice.
-	switch strings.ToLower(filepath.Base(strings.ReplaceAll(shellBin, `\`, "/"))) {
-	case "cmd", "cmd.exe":
-		return "Use native Windows commands (dir, type, copy, del, move, findstr) — NOT ls/cat/rm/cp/mv/grep. Chain with `&&`."
-	case "powershell", "powershell.exe", "pwsh", "pwsh.exe":
-		return "Use PowerShell cmdlets (Get-ChildItem, Get-Content, Copy-Item, Remove-Item) or their aliases."
-	default:
-		return "Use standard POSIX commands (ls, cat, grep, rm). Chain with `&&`."
-	}
-}
-
-// firstSentence trims a tool description to its first sentence (or first line),
-// keeping the environment preamble compact when a tool has a long description.
-func firstSentence(s string) string {
-	s = strings.TrimSpace(s)
-	if i := strings.IndexByte(s, '\n'); i >= 0 {
-		s = s[:i]
-	}
-	if i := strings.Index(s, ". "); i >= 0 {
-		s = s[:i+1]
-	}
-	return strings.TrimSpace(s)
-}
-
 // maybeForge folds the run's journal by correlation, and if the run made at
 // least SkillForgeMinTools tool calls, runs one best-effort skill proposal over
 // a compact transcript. Best-effort: any error is journaled and swallowed — a
@@ -4109,122 +2503,19 @@ func buildTranscript(toolNames []string, answer string) string {
 }
 
 // Why returns every event with the same correlation_id as the named event,
-// in seq order. It is the M0.5 form of `agt why` — useful to see what task
-// produced an event. A richer tree-form view lands later.
-func (k *Kernel) Why(eventID string) ([]*event.Event, error) {
-	var (
-		target *event.Event
-		all    []*event.Event
-	)
-	err := k.journal.Range(func(e *event.Event) error {
-		all = append(all, e)
-		if e.ID == eventID {
-			target = e
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	if target == nil {
-		return nil, fmt.Errorf("runtime: event %s not found", eventID)
-	}
-	if target.CorrelationID == "" {
-		// Singleton event with no correlation — just return it.
-		return []*event.Event{target}, nil
-	}
-	out := make([]*event.Event, 0, 16)
-	for _, e := range all {
-		if e.CorrelationID == target.CorrelationID {
-			out = append(out, e)
-		}
-	}
-	return out, nil
-}
+// in seq order (the M0.5 form of `agt why`). Delegates to the journal's
+// provenance scan (Phase 3.1 C).
+func (k *Kernel) Why(eventID string) ([]*event.Event, error) { return k.journal.Why(eventID) }
 
-// Causes returns the causation ancestry of an event: the chain of events linked
-// by causation_id from the root cause down to (and including) the target,
-// ordered oldest-first (root → … → target). This is the provenance walk
-// SPEC-01 §7.1 describes ("agt why walks the chain backwards"), and it is
-// distinct from Why, which groups by correlation_id.
-//
-// Causation crosses correlation boundaries where correlation cannot: a Pulse
-// delta/salience/initiative event carries its OWN per-chain correlation but
-// links to the originating tick (a different correlation) only via causation_id
-// — so the tick is reachable here yet invisible to Why. Likewise a channel
-// reply links to the inbound message that caused it.
-//
-// The walk is cycle-guarded (a corrupt or forged journal could contain a
-// causation loop; the runtime never emits one) and stops at the root
-// (causation_id == "") or a dangling link (the referenced parent is absent).
-// Returns the target alone when it has no causation parent.
-func (k *Kernel) Causes(eventID string) ([]*event.Event, error) {
-	byID := make(map[string]*event.Event)
-	var target *event.Event
-	err := k.journal.Range(func(e *event.Event) error {
-		byID[e.ID] = e
-		if e.ID == eventID {
-			target = e
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	if target == nil {
-		return nil, fmt.Errorf("runtime: event %s not found", eventID)
-	}
-	// Walk causation_id backwards from the target, newest-first.
-	chain := make([]*event.Event, 0, 8)
-	seen := make(map[string]struct{})
-	for cur := target; cur != nil; {
-		if _, dup := seen[cur.ID]; dup {
-			break // cycle guard: a forged causation loop must terminate
-		}
-		seen[cur.ID] = struct{}{}
-		chain = append(chain, cur)
-		if cur.CausationID == "" {
-			break // reached the root cause
-		}
-		cur = byID[cur.CausationID] // nil if the parent is missing → walk ends
-	}
-	// Reverse in place to oldest-first (root → … → target).
-	for i, j := 0, len(chain)-1; i < j; i, j = i+1, j-1 {
-		chain[i], chain[j] = chain[j], chain[i]
-	}
-	return chain, nil
-}
+// Causes returns the causation ancestry of an event, root-first — the
+// provenance walk SPEC-01 §7.1 describes. Delegates to the journal's
+// provenance scan (Phase 3.1 C).
+func (k *Kernel) Causes(eventID string) ([]*event.Event, error) { return k.journal.Causes(eventID) }
 
 // ParentOf returns the lead run's correlation for a sub-agent run, or "" if
-// childCorr was not spawned via delegation (M42). It scans the journal for a
-// subagent.spawned event whose payload names childCorr as its child — the
-// spawn lives under the PARENT correlation, so this is the only way to walk
-// child→parent (the parent→child direction is already visible because the
-// spawn is in the parent's own chain). A single forward scan; the last match
-// wins (a child has exactly one spawn in practice).
-func (k *Kernel) ParentOf(childCorr string) string {
-	if childCorr == "" {
-		return ""
-	}
-	parent := ""
-	_ = k.journal.Range(func(e *event.Event) error {
-		if e.Kind != event.KindSubAgentSpawned {
-			return nil
-		}
-		var p struct {
-			Child  string `json:"child_correlation"`
-			Parent string `json:"parent"`
-		}
-		if err := json.Unmarshal(e.Payload, &p); err != nil {
-			return nil
-		}
-		if p.Child == childCorr && p.Parent != "" {
-			parent = p.Parent
-		}
-		return nil
-	})
-	return parent
-}
+// childCorr was not spawned via delegation (M42). Delegates to the journal's
+// provenance scan (Phase 3.1 C).
+func (k *Kernel) ParentOf(childCorr string) string { return k.journal.ParentOf(childCorr) }
 
 // Verify replays every event and confirms the BLAKE3 chain is intact.
 // Returns nil on success.
